+++ sdk/keyvault/Azure.Security.KeyVault.Keys/CHANGELOG.md
@@ -1,10 +1,16 @@
 # Release History
 
-## 4.9.0-beta.1 (Unreleased)
+## 4.10.0-beta.1 (Unreleased)
 
 ### Features Added
 
 ### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 4.9.0 (2026-03-04)
 
 ### Bugs Fixed